requirements.txt
flask==3.1.0
twilio==9.4.3
openai==1.75.0
requests==2.32.3
python-dotenv==1.0.1
gunicorn==23.0.0
